refactor:#180 layout 경로 변경
