+++ src/app/(auth)/change-password/layout.tsx
@@ -1,6 +1,6 @@
-import GlobalNavigationBar from '@/components/layouts/global-navigation-bar';
-import HeaderWithProfile from '@/components/layouts/header/variants/header-with-profile';
+import Header from '@/components/layouts/header';
 import Footer from '@/components/commons/footer';
+import GlobalNavigationBar from '@/components/layouts/global-navigation-bar';
 
 const Layout = ({
   children
@@ -9,7 +9,7 @@ const Layout = ({
 }>) => {
   return (
     <div className="temp-layout">
-      <HeaderWithProfile />
+      <Header variant="withProfile" />
       <div className="p-layout">
         <main>{children}</main>
         <Footer />
