+++ src/app/(client)/set-goal/layout.tsx
@@ -1,5 +1,5 @@
 import { Suspense } from 'react';
-import Header from '@/components/layouts/header/variants/header-default';
+import Header from '@/components/layouts/header';
 
 const Layout = ({
   children
